web_gateway: carve connect_bootstrap.rs (pure move)
The Connect-relayed dashboard bootstrap moves to
web_gateway/connect_bootstrap.rs: the offer/ice/close responses for
browsers arriving via the rendezvous and the self-contained bootstrap
HTML page, plus its test.
